sound: widen s16 mixer gains to a 2.0 ceiling for SSF_UNCLAMPED parity
The short-typed Snd_ClampGainQ15 saturated all gains at unity, so an
SSF_UNCLAMPED shader mixing at 1.5 played at 1.5 through the float
pipeline and at 1.0 through the s16 pipeline. The kernels header
acknowledged the divergence and waved it off on the grounds that gains
never approached unity - true only while the removed x0.333 pad held
everything down. With the pad gone and the pre-master clamp exempting
SSF_UNCLAMPED (its authored meaning), over-unity gains are now real, so
the two pipelines' gain laws must match over the range shaders use.

Gains become int with a 2.0 ceiling: 0..65534 at the unchanged 32767 =
unity scale. 65534 is exactly the largest gain whose products with any
s16 sample stay in a single 32-bit multiply - worked at the extremes:
32767*65534 + 0x4000 = 2,147,368,962 and -32768*65534 + 0x4000 =
-2,147,401,728, both inside int32 - so the kernels keep their plain
one-multiply inner loops, no int64. Ramp intermediates: base <=
65534<<8 = 16.8M, closed-form interpolation stays between the
endpoints. The reverb send's x*sq requantizer shares the same bound;
AIRQ was already int64.

Gains beyond 2.0 still saturate (at 2.0); the float pipeline applies
them and the output knee bounds the sum. Practical shaders do not
exceed it, and the previous behavior saturated the same content at 1.0.

bench: gain endpoints widened to int; new over-unity test drives both
endpoint patterns at 65534 with alternating +-full-scale samples
against an int64 shadow reference that asserts every product fits int32
('proven') and the kernel output matches bit-exact. All tests pass
FAIL-free on scalar, SSE2, NEON/aarch64 (qemu), NEON/armv7 (qemu).
Bench binaries rebuilt.

+++ neo/sound/snd_world.cpp
@@ -1710,8 +1710,8 @@ void idSoundWorldLocal::AddChannelContribution( idSoundEmitterLocal *sound, idSo
 			const int n = stereoSample ? numFrames*2 : numFrames;
 			Snd_FloatToS16( srcS16, alignedInputSamples, n );
 			haveS16 = true;
-			short lastQ[2]    = { Snd_ClampGainQ15( chan->lastV[0] ), Snd_ClampGainQ15( chan->lastV[1] ) };
-			short currentQ[2] = { Snd_ClampGainQ15( ears[0] ),        Snd_ClampGainQ15( ears[1] ) };
+			int lastQ[2]    = { Snd_ClampGainQ15( chan->lastV[0] ), Snd_ClampGainQ15( chan->lastV[1] ) };
+			int currentQ[2] = { Snd_ClampGainQ15( ears[0] ),        Snd_ClampGainQ15( ears[1] ) };
 			int *accum = (int *)finalMixBuffer;   // s16 mode: the buffer is the int32 accumulator
 			if ( stereoSample ) {
 				Snd_MixTwoSpeakerStereoS16( accum, srcS16, numFrames, lastQ, currentQ );
@@ -1791,7 +1791,7 @@ void idSoundWorldLocal::AddChannelContribution( idSoundEmitterLocal *sound, idSo
 					}
 				}
 			} else {
-				const short sq = Snd_ClampGainQ15( sendVol );
+				const int sq = Snd_ClampGainQ15( sendVol );
 				/*
 				   srcS16 already holds this block converted, done by the mix
 				   above. The previous code declared a second buffer and ran
@@ -1815,8 +1815,8 @@ void idSoundWorldLocal::AddChannelContribution( idSoundEmitterLocal *sound, idSo
 				*/
 				#define AIRQ(x,g) ( (x) >= 0 ? (int)( ( (long long)(x) * (g) ) >> 15 ) \
 				                             : -(int)( ( -(long long)(x) * (g) ) >> 15 ) )
-				const short airAlphaQ = Snd_ClampGainQ15( airAlpha );
-				const short airHfGQ   = Snd_ClampGainQ15( airHfGain );
+				const int airAlphaQ = Snd_ClampGainQ15( airAlpha );
+				const int airHfGQ   = Snd_ClampGainQ15( airHfGain );
 				if ( stereoSample ) {
 					for ( int k = 0; k < numFrames; k++ ) {
 						int x = ( srcS16[(size_t)k*2] + srcS16[(size_t)k*2+1] ) / 2;
